% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildBranchesData.R
\name{BuildBranchesData}
\alias{BuildBranchesData}
\title{Compute default values of societal footprints by NACE branches (38)}
\usage{
BuildBranchesData(Indicator, Year)
}
\arguments{
\item{Indicator}{Considered indicator.}

\item{Year}{Considered year.}
}
\value{
A `data.frame` object containing final default values by economic activities branches.
}
\description{
Unlike BuildBranchesData function, this derived model take into
account fixed capital consumption.
}
\details{
This function returns a table summarizing final default values
of the chosen indicator on the selected year.
}
\examples{
BuildBranchesDataV2("GHG",2018)
}
\seealso{
\code{\link{BuildDivisionsData}}, \code{\link{FetchDataDisponibility}}, \code{\link{BuildBranchesData}}.
}
