% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGEQData.R
\name{BuildGEQData}
\alias{BuildGEQData}
\title{Build and returns all data required to the GEQ indicator computations.}
\usage{
BuildGEQData(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
An object `list` made up of 4 elements : value added impacts by French branches,
imported products associated coefficient, data sources and values unit.
}
\description{
Returns a `list` made up of value added impacts by French branches, imported products associated coefficient,
Data sources and values unit. This data will be used in both BuildBranchesData and BuildDivisionsData functions.
}
\examples{
BuildGEQData(max(FetchDataDisponibility("GEQ"))
}
\seealso{
\code{\link{BuildECOData}}, \code{\link{BuildGHGData}}, \code{\link{BuildNRGData}},
 \code{\link{BuildBranchesData}}, \code{\link{BuildDivisionsData}}, \code{\link{FetchDataDisponibility}}.
}
