% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGHGData.R
\name{BuildGHGData}
\alias{BuildGHGData}
\title{Build and returns all data required to the GHG indicator computations.}
\usage{
BuildGHGData(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
An object `list` made up of 5 elements : value added impacts by French branches,
imported products associated coefficient, data sources, values unit and value added impacts by French divisions.
}
\description{
Returns a `list` made up of value added impacts by French branches, imported products associated coefficient,
Data sources, values unit and value added impacts by French divisions. This data will be used in both BuildBranchesData and BuildDivisionsData functions.
}
\examples{
BuildGHGData(max(FetchDataDisponibility("GHG"))
}
\seealso{
\code{\link{BuildECOData}}, \code{\link{BuildNRGData}},
 \code{\link{BuildBranchesData}}, \code{\link{BuildDivisionsData}}, \code{\link{FetchDataDisponibility}}.
}
