% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildKNWData.R
\name{BuildKNWData}
\alias{BuildKNWData}
\title{Build and returns all data required to the KNW indicator computations.}
\usage{
BuildKNWData(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
An object `list` made up of 4 elements : value added impacts by French branches,
imported products associated coefficient, data sources and values unit.
}
\description{
Returns a `list` made up of value added impacts by French branches, imported products associated coefficient,
Data sources and values unit. This data will be used in both BuildBranchesData and BuildDivisionsData functions.
}
\examples{
BuildKNWData(max(FetchDataDisponibility("KNW"))
}
\seealso{
\code{\link{BuildECOData}}, \code{\link{BuildGHGData}},
 \code{\link{BuildBranchesData}}, \code{\link{BuildDivisionsData}}, \code{\link{FetchDataDisponibility}}.
}
