% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSOCData.R
\name{BuildSOCData}
\alias{BuildSOCData}
\title{Build and returns all data required to the SOC indicator computations.}
\usage{
BuildSOCData(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
An object `list` made up of 4 elements : value added impacts by French branches,
imported products associated coefficient, data sources and values unit.
}
\description{
Returns a `list` made up of value added impacts by French branches, imported products associated coefficient,
Data sources and values unit. This data will be used in both BuildBranchesData and BuildDivisionsData functions.
}
\examples{
BuildSOCData(max(FetchDataDisponibility("SOC"))
}
\seealso{
\code{\link{BuildECOData}}, \code{\link{BuildGHGData}},
 \code{\link{BuildBranchesData}}, \code{\link{BuildDivisionsData}}, \code{\link{FetchDataDisponibility}}.
}
