% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deflator.R
\name{Deflator}
\alias{Deflator}
\title{Correct monetary value with respect to French inflation.}
\usage{
Deflator(Value, Start, Target)
}
\arguments{
\item{Value}{Initial value to deflate.}

\item{Start}{Initial y.}

\item{Target}{Initial value to deflate.}
}
\value{
A `numeric` value.
}
\description{
Returns a deflated value.
}
\examples{
Deflator(1250,1990,2020)
}
\seealso{
\code{\link{BuildBranchesData}}, \code{\link{BuildDivisionsData}}, \code{\link{FetchDataDisponibility}}.
}
