% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchDataCFC.R
\name{FetchDataCFC}
\alias{FetchDataCFC}
\title{Fetch data frame associated to branches consumption of fixed capital
of annual national accounts from Insee's API.}
\usage{
FetchDataCFC(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
A `data.frame` object containing annual Insee supply-use equilibrium data.
}
\description{
Fetch data frame associated to branches consumption of fixed capital
of annual national accounts from Insee's API.
}
\examples{
FetchDataCCF(2018)
}
\seealso{
\code{\link{BuildDivisionsData}}, \code{\link{BuildBranchesData}}.
}
