% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchDataCPEB.R
\name{FetchDataCPEB}
\alias{FetchDataCPEB}
\title{Fetch data frame associated to production and business accounts
of annual national accounts from Insee's API.}
\usage{
FetchDataCPEB(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
A `data.frame` object containing annual Insee production and business accounts data.
}
\description{
Fetch data frame associated to production and business accounts
of annual national accounts from Insee's API.
}
\examples{
FetchDataCPEB(2018)
}
\seealso{
\code{\link{BuildDivisionsData}}, \code{\link{BuildBranchesData}}.
}
