% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchDataDisponibility.R
\name{FetchDataDisponibility}
\alias{FetchDataDisponibility}
\title{Fetch societal data disponibility.}
\usage{
FetchDataDisponibility(Indicator)
}
\arguments{
\item{Indicator}{Considered year.}
}
\value{
An `integer` object containing available year(s) for considered indicator.
}
\description{
Returns indicator-specific available years, in line with data sources used. An automation of the function actualization process is in working progress.
}
\examples{
Year=FetchDataDisponibility("GHG")
BuildGHGData(Year)
}
\seealso{
\code{\link{BuildECOData}}, \code{\link{BuildGHGData}}, \code{\link{BuildNRGData}},
 \code{\link{BuildBranchesData}}, \code{\link{BuildDivisionsData}}, \code{\link{FetchDataDisponibility}}.
}
