% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchDataERE.R
\name{FetchDataERE}
\alias{FetchDataERE}
\title{Fetch data frame associated to supply-use equilibrium
of annual national accounts from Insee's API.}
\usage{
FetchDataERE(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
A `data.frame` object containing annual Insee supply-use equilibrium data.
}
\description{
Fetch data frame associated to supply-use equilibrium
of annual national accounts from Insee's API.
}
\examples{
FetchDataERE(2018)
}
\seealso{
\code{\link{BuildDivisionsData}}, \code{\link{BuildBranchesData}}.
}
