% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchDataTEI.R
\name{FetchDataTEI}
\alias{FetchDataTEI}
\title{Fetch data frame associated to intermediate use tables
of annual national accounts from Insee's API.}
\usage{
FetchDataTEI(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
A `data.frame` object containing annual Insee intermediate use tables data.
}
\description{
Fetch data frame associated to intermediate use tables
of annual national accounts from Insee's API.
}
\examples{
FetchDataTEI(2018)
}
\seealso{
\code{\link{BuildDivisionsData}}, \code{\link{BuildBranchesData}}.
}
