% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchDataTESS.R
\name{FetchDataTESS}
\alias{FetchDataTESS}
\title{Fetch data frame associated to domestic symetric input-output tables
of Insee annual national accounts.}
\usage{
FetchDataTESS(Year)
}
\arguments{
\item{Year}{Considered year.}
}
\value{
A `data.frame` object containing annual Insee domestic symetric input-output tables data.
}
\description{
Fetch data frame associated to domestic symetric input-output tables
of Insee annual national accounts.
}
\examples{
FetchDataTESS(2018)
}
\seealso{
\code{\link{BuildDivisionsData}}, \code{\link{BuildBranchesData}}.
}
