% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndicatorList.R
\name{get_indicator_list}
\alias{get_indicator_list}
\title{Get non-financial data list and main informations.}
\usage{
get_indicator_list()
}
\value{
A table of all 12 supplied non-financial dimensions.
}
\description{
Get non-financial data list and main informations.
}
\details{
Get non-financial data list and main informations : full label, code, unit, data sources and used rounding.
}
\examples{
getIndicatorList()

}
\seealso{
\code{\link{build_branches_fpt}}, \code{\link{build_divisions_fpt}}, \code{\link{get_indicator_list}}
}
