% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnr-package.R
\docType{package}
\name{lsnr-package}
\alias{lsnr}
\alias{lsnr-package}
\title{lsnr: Societal footprint by economic activities}
\description{
This package provides a set of R functions designed to compute French economic activities societal footprints. More precisely, they allow the user to roughly modelize French economic activity and its interactions and to compute related societal footprints by 38 NACE branches and 88 NACE divisions. It offers a 12 non-financial dimensions framework. Detailed documentation and metholodology are availables online \url{https://docs.lasocietenouvelle.org/}. This package is designed by La 'Societe Nouvelle', a French organization that set up a non financial national information system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/La-Societe-Nouvelle}
  \item \url{https://lasocietenouvelle.org/}
  \item \url{https://lasocietenouvelle.org/portail}
  \item \url{https://www.linkedin.com/company/la-societe-nouvelle/.}
}

}
\author{
\strong{Maintainer}: Joris BLAIN \email{joris.blain@lasocietenouvelle.org}

}
\keyword{internal}
