% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDivisionsData.R
\name{build_divisions_fpt}
\alias{build_divisions_fpt}
\title{Compute macroeconomic values of societal footprints by NACE divisions (88)}
\usage{
build_divisions_fpt(indicator, year, verbose = T)
}
\arguments{
\item{indicator}{requested non-financial dimension.}

\item{year}{year of requested data.}
}
\value{
A table of macroeconomic footprint values by economic activities division.
}
\description{
Compute macroeconomic values of societal footprints by NACE divisions (88)
}
\details{
This function aims to compute French branch societal footprints by non-financial dimension by year.
It involves, on one hand, a macroeconomic input-output modelization of the French economy and its interactions with
the rest of the world, based on INSEE IOTs, and requires, on the other hand, direct impact data from institutional sources.
Specifically BuildDivisionsData compute division footprints by adjusting branch footprints thanks to
division economic aggregate values.
}
\examples{
build_divisions_fpt("ECO",2018)
}
\seealso{
\code{\link{build_branches_fpt}}, \code{\link{get_indicator_list}}
}
